implement new use case just for getting records
