+++ SocialCareCaseViewerApi.Tests/V1/UseCase/ProcessDataUseCaseTests.cs
@@ -18,16 +18,14 @@ namespace SocialCareCaseViewerApi.Tests.V1.UseCase
     public class ProcessDataUseCaseTests
     {
         private Mock<IProcessDataGateway> _mockProcessDataGateway;
-        private Mock<IDatabaseGateway> _mockDatabaseGateway;
         private ProcessDataUseCase _classUnderTest;
-        private Fixture _fixture = new Fixture();
+        private readonly Fixture _fixture = new Fixture();
 
         [SetUp]
         public void SetUp()
         {
             _mockProcessDataGateway = new Mock<IProcessDataGateway>();
-            _mockDatabaseGateway = new Mock<IDatabaseGateway>();
-            _classUnderTest = new ProcessDataUseCase(_mockProcessDataGateway.Object, _mockDatabaseGateway.Object);
+            _classUnderTest = new ProcessDataUseCase(_mockProcessDataGateway.Object);
         }
 
         // [Test]
