+++ SocialCareCaseViewerApi/V1/UseCase/Interfaces/IGetRecordsUseCase.cs
@@ -1,0 +1,10 @@
+using SocialCareCaseViewerApi.V1.Boundary.Requests;
+using SocialCareCaseViewerApi.V1.Boundary.Response;
+
+namespace SocialCareCaseViewerApi.V1.UseCase.Interfaces
+{
+    public interface IGetRecordsUseCase
+    {
+        CareCaseDataList Execute(GetRecordsRequest request);
+    }
+}
